#COBUMMFO-10035# Коментар: MERGE: COBUMMFO-10035 Анна Крахмальова

+++ ABS/master/web/mmfo/barsroot/App_Code/Areas/Payreg/Models/RegularSbonWithContractOrder.cs
@@ -5,7 +5,7 @@ namespace BarsWeb.Areas.Payreg.Models
     public class RegularSbonWithContractOrder
     {
         public int? Id { get; set; }
-        public int PayerAccountId { get; set; }
+        public long PayerAccountId { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime? StopDate { get; set; }
         public int PaymentFrequency { get; set; }
